Moved Frameless behavior to mixin
